Minor type change.

+++ contracts/RewardCalculator.sol
@@ -3,7 +3,7 @@ pragma solidity ^0.4.23;
 
 contract RewardCalculator {
 
-    function calculateRewardAmount(uint8 rating, uint64 price)
+    function calculateRewardAmount(uint8 rating, uint price)
     external pure returns(uint RewardAmount) {
         RewardAmount = rating * price; // This is preliminary.
     }
